--- a/log-viewer/modules/parsers/ApexLogParser.ts
+++ b/log-viewer/modules/parsers/ApexLogParser.ts
@@ -681,9 +681,10 @@ export abstract class LogLine {
 
   constructor(parts: string[] | null) {
     if (parts) {
-      this.type = parts[1] as LogEventType;
+      const [timeData, type] = parts;
+      this.type = type as LogEventType;
       this.text = this.type;
-      this.timestamp = parseTimestamp(parts[0] || '');
+      this.timestamp = this.parseTimestamp(timeData as string);
     }
   }
 
@@ -692,6 +693,30 @@ export abstract class LogLine {
 
   /** Called when the Log event after this one is created in the line parser*/
   onAfter?(parser: ApexLogParser, next?: LogLine): void;
+
+  private parseTimestamp(text: string): number {
+    const start = text.indexOf('(');
+    if (start !== -1) {
+      return Number(text.slice(start + 1, -1));
+    }
+    throw new Error(`Unable to parse timestamp: '${text}'`);
+  }
+
+  protected parseLineNumber(text: string | null | undefined): string | number {
+    switch (true) {
+      case text === '[EXTERNAL]':
+        return 'EXTERNAL';
+      case !!text: {
+        const lineNumberStr = text.slice(1, -1);
+        if (lineNumberStr) {
+          return Number(lineNumberStr);
+        }
+        throw new Error(`Unable to parse line number: '${text}'`);
+      }
+      default:
+        return 0;
+    }
+  }
 }
 
 class BasicLogLine extends LogLine {}
@@ -862,30 +887,6 @@ export function parseVfNamespace(text: string): string {
   return text.substring(secondSlash + 1, sep);
 }
 
-export function parseTimestamp(text: string): number {
-  const start = text.indexOf('(');
-  if (start !== -1) {
-    return Number(text.slice(start + 1, -1));
-  }
-  throw new Error(`Unable to parse timestamp: '${text}'`);
-}
-
-export function parseLineNumber(text: string | null | undefined): string | number {
-  switch (true) {
-    case text === '[EXTERNAL]':
-      return 'EXTERNAL';
-    case !!text: {
-      const lineNumberStr = text.slice(1, -1);
-      if (lineNumberStr) {
-        return Number(lineNumberStr);
-      }
-      throw new Error(`Unable to parse line number: '${text}'`);
-    }
-    default:
-      return 0;
-  }
-}
-
 export function parseRows(text: string | null | undefined): number {
   if (!text) {
     return 0;
@@ -949,7 +950,7 @@ class ConstructorEntryLine extends Method {
 
   constructor(parts: string[]) {
     super(parts, ['CONSTRUCTOR_EXIT'], 'Method', 'method');
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     const args = parts[4];
 
     this.text = parts[5] + (args ? args.substring(args.lastIndexOf('(')) : '');
@@ -961,7 +962,7 @@ class ConstructorExitLine extends LogLine {
 
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
   }
 }
 
@@ -969,7 +970,7 @@ class EmailQueueLine extends LogLine {
   acceptsText = true;
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
   }
 }
 
@@ -978,7 +979,7 @@ export class MethodEntryLine extends Method {
 
   constructor(parts: string[]) {
     super(parts, ['METHOD_EXIT'], 'Method', 'method');
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = parts[4] || this.type || '';
     if (this.text === 'System.Type.forName(String, String)') {
       this.cpuType = 'loading'; // assume we are not charged for class loading (or at least not lengthy remote-loading / compiling)
@@ -991,7 +992,7 @@ class MethodExitLine extends LogLine {
 
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
   }
 }
 
@@ -1000,7 +1001,7 @@ class SystemConstructorEntryLine extends Method {
 
   constructor(parts: string[]) {
     super(parts, ['SYSTEM_CONSTRUCTOR_EXIT'], 'System Method', 'method');
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = parts[3] || '';
   }
 }
@@ -1010,13 +1011,13 @@ class SystemConstructorExitLine extends LogLine {
 
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
   }
 }
 class SystemMethodEntryLine extends Method {
   constructor(parts: string[]) {
     super(parts, ['SYSTEM_METHOD_EXIT'], 'System Method', 'method');
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = parts[3] || '';
   }
 }
@@ -1026,7 +1027,7 @@ class SystemMethodExitLine extends LogLine {
 
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
   }
 }
 
@@ -1108,7 +1109,7 @@ class VFApexCallStartLine extends Method {
 
   constructor(parts: string[]) {
     super(parts, ['VF_APEX_CALL_END'], 'Method', 'method');
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
 
     const classText = parts[5] || parts[3] || '';
     let methodtext = parts[4] || '';
@@ -1202,7 +1203,7 @@ class DMLBeginLine extends Method {
 
   constructor(parts: string[]) {
     super(parts, ['DML_END'], 'DML', 'free');
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = 'DML ' + parts[3] + ' ' + parts[4];
     const rowCountString = parts[5];
     this.rowCount.total = this.rowCount.self = rowCountString ? parseRows(rowCountString) : 0;
@@ -1214,14 +1215,14 @@ class DMLEndLine extends LogLine {
 
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
   }
 }
 
 class IdeasQueryExecuteLine extends LogLine {
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
   }
 }
 
@@ -1234,7 +1235,7 @@ class SOQLExecuteBeginLine extends Method {
 
   constructor(parts: string[]) {
     super(parts, ['SOQL_EXECUTE_END'], 'SOQL', 'free');
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
 
     const [, , , aggregations, soqlString] = parts;
 
@@ -1256,7 +1257,7 @@ class SOQLExecuteEndLine extends LogLine {
 
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.rowCount.total = this.rowCount.self = parseRows(parts[3] || '');
   }
 }
@@ -1271,7 +1272,7 @@ class SOQLExecuteExplainLine extends LogLine {
 
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
 
     const queryPlanDetails = parts[3] || '';
     this.text = queryPlanDetails;
@@ -1310,7 +1311,7 @@ class SOQLExecuteExplainLine extends LogLine {
 class SOSLExecuteBeginLine extends Method {
   constructor(parts: string[]) {
     super(parts, ['SOSL_EXECUTE_END'], 'SOQL', 'free');
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = `SOSL: ${parts[3]}`;
   }
 
@@ -1324,7 +1325,7 @@ class SOSLExecuteEndLine extends LogLine {
 
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.rowCount.total = this.rowCount.self = parseRows(parts[3] || '');
   }
 }
@@ -1332,7 +1333,7 @@ class SOSLExecuteEndLine extends LogLine {
 class HeapAllocateLine extends LogLine {
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = parts[3] || '';
   }
 }
@@ -1340,21 +1341,21 @@ class HeapAllocateLine extends LogLine {
 class HeapDeallocateLine extends LogLine {
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
   }
 }
 
 class StatementExecuteLine extends LogLine {
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
   }
 }
 
 class VariableScopeBeginLine extends LogLine {
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = parts.slice(3).join(' | ');
   }
 }
@@ -1362,14 +1363,14 @@ class VariableScopeBeginLine extends LogLine {
 class VariableAssignmentLine extends LogLine {
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = parts.slice(3).join(' | ');
   }
 }
 class UserInfoLine extends LogLine {
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = parts[3] + ' ' + parts[4];
   }
 }
@@ -1379,7 +1380,7 @@ class UserDebugLine extends LogLine {
 
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = parts.slice(3).join(' | ');
   }
 }
@@ -1407,7 +1408,7 @@ class CumulativeProfilingBeginLine extends Method {
 class LimitUsageLine extends LogLine {
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = parts[3] + ' ' + parts[4] + ' out of ' + parts[5];
   }
 }
@@ -1488,7 +1489,7 @@ class PushTraceFlagsLine extends LogLine {
 
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = parts[4] + ', line:' + this.lineNumber + ' - ' + parts[5];
   }
 }
@@ -1498,7 +1499,7 @@ class PopTraceFlagsLine extends LogLine {
 
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = parts[4] + ', line:' + this.lineNumber + ' - ' + parts[5];
   }
 }
@@ -1506,7 +1507,7 @@ class PopTraceFlagsLine extends LogLine {
 class QueryMoreBeginLine extends Method {
   constructor(parts: string[]) {
     super(parts, ['QUERY_MORE_END'], 'SOQL', 'custom');
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = `line: ${this.lineNumber}`;
   }
 }
@@ -1516,14 +1517,14 @@ class QueryMoreEndLine extends LogLine {
 
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = `line: ${this.lineNumber}`;
   }
 }
 class QueryMoreIterationsLine extends LogLine {
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = `line: ${this.lineNumber}, iterations:${parts[3]}`;
   }
 }
@@ -1531,7 +1532,7 @@ class QueryMoreIterationsLine extends LogLine {
 class SavepointRollbackLine extends LogLine {
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = `${parts[3]}, line: ${this.lineNumber}`;
   }
 }
@@ -1539,7 +1540,7 @@ class SavepointRollbackLine extends LogLine {
 class SavePointSetLine extends LogLine {
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = `${parts[3]}, line: ${this.lineNumber}`;
   }
 }
@@ -2270,7 +2271,7 @@ class ExceptionThrownLine extends LogLine {
 
   constructor(parts: string[]) {
     super(parts);
-    this.lineNumber = parseLineNumber(parts[2]);
+    this.lineNumber = this.parseLineNumber(parts[2]);
     this.text = parts[3] || '';
   }