# Description

- Generally improves parser performance.
- The main change is to remove recursion from the aggregate totals function
- Add a fast path for getting commonly used event types
- Add a fast path for splitting new lines when the line endings are all lf e.g `\n`

Total parse time: from 2.02s -> 1.95s 
toTree: ~1022ms -> 1025 ms
parseLog: ~974-> 750 ms


- [ ] 🐛 Bug fix
- [ ] ✨ New feature
- [ ] ♻️ Refactor
- [X] ⚡ Performance Improvement
- [ ] 📝 Documentation
- [ ] 🔧 Chore
- [ ] 💥 Breaking change

## [optional] Any images / gifs / video

## Related Tickets & Documents

Related Issue #
fixes #
resolves #
closes #

## Added tests?

- [X] 👍 yes
- [ ] 🙅 no, not needed
- [ ] 🙋 no, I need help

## Added to documentation?

- [ ] 🔖 README.md
- [ ] 🔖 CHANGELOG.md
- [ ] 📖 help site
- [X ] 🙅 not needed

## [optional] Are there any post-deployment tasks we need to perform?